requirements.txt
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.101.1
greenlet==3.1.1
h11==0.14.0
httptools==0.6.1
idna==3.10
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.23.2
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1

passlib~=1.7.4
python-jose~=3.3.0
asyncio~=3.4.3
pytest~=8.3.3
httpx~=0.27.2
alembic~=1.13.3